requirements.txt
asgiref==3.4.1
automium==0.2.6
automium-web==0.1.1
click==8.0.3
Django==3.2.9
django-extensions==3.1.5
factory-boy==3.2.1
Faker==9.9.0
flake8==4.0.1
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-plugin-utils==1.3.2
flake8-print==4.0.0
flake8-quotes==3.3.1
flake8-requirements==1.5.1
flake8-return==1.1.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
legacy==0.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
Werkzeug==2.0.2
